# docs/api-allergy-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RxFlow - 过敏档案管理",
    "version": "1.0.0",
    "description": "患者过敏档案 CRUD 接口 + 下拉数据接口"
  },
  "servers": [
    { "url": "http://192.168.1.8:3001", "description": "开发环境" },
    { "url": "http://localhost:3001", "description": "本地环境" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/patient-allergies": {
      "get": {
        "tags": ["过敏档案"],
        "summary": "查询过敏档案列表",
        "parameters": [
          { "name": "patientName", "in": "query", "description": "患者姓名模糊搜索", "schema": { "type": "string" } },
          { "name": "allergenName", "in": "query", "description": "过敏原名称模糊搜索", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "description": "页码，默认1", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "description": "每页条数，默认20", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AllergyListResponse" } } } }
        }
      },
      "post": {
        "tags": ["过敏档案"],
        "summary": "创建过敏档案",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AllergyCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "创建成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PatientAllergy" } } } }
        }
      }
    },
    "/api/patient-allergies/{id}": {
      "get": {
        "tags": ["过敏档案"],
        "summary": "查询过敏档案详情",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PatientAllergy" } } } }
        }
      },
      "put": {
        "tags": ["过敏档案"],
        "summary": "更新过敏档案",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AllergyUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "更新成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PatientAllergy" } } } }
        }
      },
      "delete": {
        "tags": ["过敏档案"],
        "summary": "删除过敏档案",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "删除成功" }
        }
      }
    },
    "/api/allergens": {
      "get": {
        "tags": ["下拉数据"],
        "summary": "获取过敏原列表（表单下拉用）",
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AllergenListResponse" } } } }
        }
      }
    },
    "/api/patients": {
      "get": {
        "tags": ["下拉数据"],
        "summary": "获取患者列表（表单下拉用）",
        "parameters": [
          { "name": "pageSize", "in": "query", "description": "返回条数", "schema": { "type": "integer", "default": 100 } }
        ],
        "responses": {
          "200": { "description": "成功", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/PatientOption" } } } } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "PatientAllergy": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "patientId": { "type": "integer", "example": 1 },
          "allergenId": { "type": "integer", "example": 1 },
          "severity": { "type": "string", "enum": ["mild", "moderate", "severe"], "description": "mild=轻度, moderate=中度, severe=重度", "example": "severe" },
          "remark": { "type": "string", "example": "曾出现过过敏性休克" },
          "source": { "type": "string", "enum": ["manual", "migration"], "description": "manual=人工录入, migration=系统迁移", "example": "migration" },
          "createdAt": { "type": "string", "format": "date-time" },
          "patient": { "$ref": "#/components/schemas/PatientBrief" },
          "allergen": { "$ref": "#/components/schemas/AllergenBrief" }
        }
      },
      "AllergyListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/PatientAllergy" } },
          "total": { "type": "integer", "example": 2 },
          "page": { "type": "integer", "example": 1 },
          "pageSize": { "type": "integer", "example": 20 }
        }
      },
      "AllergyCreateRequest": {
        "type": "object",
        "required": ["patientId", "allergenId"],
        "properties": {
          "patientId": { "type": "integer", "description": "患者ID", "example": 1 },
          "allergenId": { "type": "integer", "description": "过敏原ID", "example": 1 },
          "severity": { "type": "string", "enum": ["mild", "moderate", "severe"], "default": "moderate", "example": "moderate" },
          "remark": { "type": "string", "description": "备注", "example": "服用后全身皮疹" }
        }
      },
      "AllergyUpdateRequest": {
        "type": "object",
        "properties": {
          "patientId": { "type": "integer", "description": "患者ID" },
          "allergenId": { "type": "integer", "description": "过敏原ID" },
          "severity": { "type": "string", "enum": ["mild", "moderate", "severe"] },
          "remark": { "type": "string" }
        }
      },
      "PatientBrief": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "example": "赵建国" },
          "phone": { "type": "string", "example": "13800004444" }
        }
      },
      "AllergenBrief": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "example": "青霉素" },
          "category": { "type": "string", "enum": ["antibiotic", "antipyretic", "enzyme", "other"], "description": "antibiotic=抗生素, antipyretic=解热镇痛, enzyme=酶类, other=其他" }
        }
      },
      "AllergenListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/AllergenBrief" } }
        }
      },
      "PatientOption": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "phone": { "type": "string" }
        }
      }
    }
  }
}